feat(tech-insights): move client into separate package and allow using backend auth
