+++ workspaces/tech-insights/.changeset/lemon-months-divide.md
@@ -1,0 +1,5 @@
+---
+'@backstage-community/plugin-tech-insights': patch
+---
+
+Move TechInsightsClient and in a separate package and allow using the backend AuthService
