+++ workspaces/tech-insights/packages/backend/package.json
@@ -23,9 +23,13 @@
   "dependencies": {
     "@backstage-community/plugin-tech-insights-backend": "workspace:^",
     "@backstage-community/plugin-tech-insights-backend-module-jsonfc": "workspace:^",
+    "@backstage-community/plugin-tech-insights-node": "workspace:^",
     "@backstage/backend-common": "^0.22.0",
     "@backstage/backend-defaults": "^0.2.18",
+    "@backstage/backend-plugin-api": "^0.6.18",
     "@backstage/backend-tasks": "^0.5.23",
+    "@backstage/catalog-client": "^1.6.5",
+    "@backstage/catalog-model": "^1.5.0",
     "@backstage/config": "^1.2.0",
     "@backstage/plugin-app-backend": "^0.3.67",
     "@backstage/plugin-auth-backend": "^0.22.5",
