+++ workspaces/tech-insights/plugins/tech-insights-client/.eslintrc.js
@@ -1,0 +1,1 @@
+module.exports = require('@backstage/cli/config/eslint-factory')(__dirname);
