+++ workspaces/tech-insights/plugins/tech-insights-client/README.md
@@ -1,0 +1,12 @@
+# @backstage-community/tech-insights
+
+_This package was created through the Backstage CLI_.
+
+## Installation
+
+Install the package via Yarn:
+
+```sh
+cd <package-dir> # if within a monorepo
+yarn add @backstage-community/tech-insights
+```
