+++ workspaces/tech-insights/plugins/tech-insights-client/api-report.md.api.md
@@ -1,0 +1,71 @@
+## API Report File for "@backstage-community/plugin-tech-insights-client"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { AuthService } from '@backstage/backend-plugin-api';
+import { BulkCheckResponse } from '@backstage-community/plugin-tech-insights-common';
+import { CheckResult } from '@backstage-community/plugin-tech-insights-common';
+import { CompoundEntityRef } from '@backstage/catalog-model';
+import { DiscoveryApi } from '@backstage/core-plugin-api';
+import { FactSchema } from '@backstage-community/plugin-tech-insights-common';
+import { IdentityApi } from '@backstage/core-plugin-api';
+import { JsonValue } from '@backstage/types';
+
+// @public
+export type Check = {
+  id: string;
+  type: string;
+  name: string;
+  description: string;
+  factIds: string[];
+  successMetadata?: Record<string, unknown>;
+  failureMetadata?: Record<string, unknown>;
+};
+
+// @public
+export interface InsightFacts {
+  // (undocumented)
+  [factId: string]: {
+    timestamp: string;
+    version: string;
+    facts: Record<string, JsonValue>;
+  };
+}
+
+// @public (undocumented)
+export class TechInsightsClient {
+  constructor(options: {
+    discoveryApi: DiscoveryApi;
+    identityApi: IdentityApi | AuthService;
+  });
+  // (undocumented)
+  getAllChecks(): Promise<Check[]>;
+  // (undocumented)
+  getFacts(entity: CompoundEntityRef, facts: string[]): Promise<InsightFacts>;
+  // (undocumented)
+  getFactSchemas(): Promise<FactSchema[]>;
+  // (undocumented)
+  runBulkChecks(
+    entities: CompoundEntityRef[],
+    checks?: Check[],
+  ): Promise<BulkCheckResponse>;
+  // (undocumented)
+  runChecks(
+    entityParams: CompoundEntityRef,
+    checks?: string[],
+  ): Promise<CheckResult[]>;
+}
+
+// Warnings were encountered during analysis:
+//
+// src/TechInsightsClient.d.ts:7:1 - (ae-undocumented) Missing documentation for "TechInsightsClient".
+// src/TechInsightsClient.d.ts:14:5 - (ae-undocumented) Missing documentation for "getFacts".
+// src/TechInsightsClient.d.ts:15:5 - (ae-undocumented) Missing documentation for "getAllChecks".
+// src/TechInsightsClient.d.ts:16:5 - (ae-undocumented) Missing documentation for "getFactSchemas".
+// src/TechInsightsClient.d.ts:17:5 - (ae-undocumented) Missing documentation for "runChecks".
+// src/TechInsightsClient.d.ts:18:5 - (ae-undocumented) Missing documentation for "runBulkChecks".
+// src/types.d.ts:51:5 - (ae-undocumented) Missing documentation for "__index".
+
+// (No @packageDocumentation comment for this package)
+```
