+++ workspaces/tech-insights/plugins/tech-insights-client/package.json
@@ -1,0 +1,44 @@
+{
+  "name": "@backstage-community/plugin-tech-insights-client",
+  "version": "0.1.0",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "license": "Apache-2.0",
+  "publishConfig": {
+    "access": "public",
+    "main": "dist/index.cjs.js",
+    "types": "dist/index.d.ts"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/backstage/community-plugins",
+    "directory": "workspaces/tech-insights/plugins/tech-insights-client"
+  },
+  "backstage": {
+    "role": "node-library"
+  },
+  "scripts": {
+    "start": "backstage-cli package start",
+    "build": "backstage-cli package build",
+    "lint": "backstage-cli package lint",
+    "test": "backstage-cli package test",
+    "clean": "backstage-cli package clean",
+    "prepack": "backstage-cli package prepack",
+    "postpack": "backstage-cli package postpack"
+  },
+  "dependencies": {
+    "@backstage-community/plugin-tech-insights-common": "workspace:^",
+    "@backstage/backend-plugin-api": "^0.6.18",
+    "@backstage/catalog-model": "^1.5.0",
+    "@backstage/core-plugin-api": "^1.9.2",
+    "@backstage/errors": "^1.2.4",
+    "@backstage/types": "^1.1.1",
+    "qs": "^6.9.4"
+  },
+  "devDependencies": {
+    "@backstage/cli": "^0.26.6"
+  },
+  "files": [
+    "dist"
+  ]
+}
