+++ workspaces/tech-insights/plugins/tech-insights-client/src/TechInsightsClient.ts
@@ -1,0 +1,134 @@
+/*
+ * Copyright 2021 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {
+  BulkCheckResponse,
+  CheckResult,
+  FactSchema,
+} from '@backstage-community/plugin-tech-insights-common';
+import { Check, InsightFacts } from './types';
+import { DiscoveryApi, IdentityApi } from '@backstage/core-plugin-api';
+import { ResponseError } from '@backstage/errors';
+import {
+  CompoundEntityRef,
+  stringifyEntityRef,
+} from '@backstage/catalog-model';
+import qs from 'qs';
+import { AuthService } from '@backstage/backend-plugin-api';
+
+/** @public */
+export class TechInsightsClient {
+  private readonly discoveryApi: DiscoveryApi;
+  private readonly identityApi: IdentityApi | AuthService;
+
+  constructor(options: {
+    discoveryApi: DiscoveryApi;
+    identityApi: IdentityApi | AuthService;
+  }) {
+    this.discoveryApi = options.discoveryApi;
+    this.identityApi = options.identityApi;
+  }
+
+  async getFacts(
+    entity: CompoundEntityRef,
+    facts: string[],
+  ): Promise<InsightFacts> {
+    const query = qs.stringify({
+      entity: stringifyEntityRef(entity),
+      ids: facts,
+    });
+    return await this.api<InsightFacts>(`/facts/latest?${query}`);
+  }
+
+  async getAllChecks(): Promise<Check[]> {
+    return this.api('/checks');
+  }
+
+  async getFactSchemas(): Promise<FactSchema[]> {
+    return this.api('/fact-schemas');
+  }
+
+  async runChecks(
+    entityParams: CompoundEntityRef,
+    checks?: string[],
+  ): Promise<CheckResult[]> {
+    const { namespace, kind, name } = entityParams;
+    const requestBody = { checks };
+    return this.api(
+      `/checks/run/${encodeURIComponent(namespace)}/${encodeURIComponent(
+        kind,
+      )}/${encodeURIComponent(name)}`,
+      {
+        method: 'POST',
+        body: JSON.stringify(requestBody),
+      },
+    );
+  }
+
+  async runBulkChecks(
+    entities: CompoundEntityRef[],
+    checks?: Check[],
+  ): Promise<BulkCheckResponse> {
+    const checkIds = checks ? checks.map(check => check.id) : [];
+    const requestBody = {
+      entities,
+      checks: checkIds.length > 0 ? checkIds : undefined,
+    };
+    return this.api('/checks/run', {
+      method: 'POST',
+      body: JSON.stringify(requestBody),
+    });
+  }
+
+  private async api<T>(path: string, init?: RequestInit): Promise<T> {
+    const url = await this.discoveryApi.getBaseUrl('tech-insights');
+    const token = await this.getToken();
+
+    const headers: HeadersInit = new Headers(init?.headers);
+    if (!headers.has('content-type'))
+      headers.set('content-type', 'application/json');
+    if (token && !headers.has('authorization')) {
+      headers.set('authorization', `Bearer ${token}`);
+    }
+
+    const request = new Request(`${url}${path}`, {
+      ...init,
+      headers,
+    });
+
+    return fetch(request).then(async response => {
+      if (!response.ok) {
+        throw await ResponseError.fromResponse(response);
+      }
+      return response.json() as Promise<T>;
+    });
+  }
+
+  private async getToken(): Promise<string | null> {
+    let result: { token?: string | undefined };
+
+    if ('getCredentials' in this.identityApi) {
+      result = await this.identityApi.getCredentials();
+    } else {
+      result = await this.identityApi.getPluginRequestToken({
+        onBehalfOf: await this.identityApi.getOwnServiceCredentials(),
+        targetPluginId: 'tech-insights',
+      });
+    }
+
+    return result.token ?? null;
+  }
+}
