+++ workspaces/tech-insights/plugins/tech-insights-client/src/setupTests.ts
@@ -1,0 +1,1 @@
+export {};
