+++ workspaces/tech-insights/plugins/tech-insights/package.json
@@ -32,6 +32,7 @@
     "test": "backstage-cli package test"
   },
   "dependencies": {
+    "@backstage-community/plugin-tech-insights-client": "workspace:^",
     "@backstage-community/plugin-tech-insights-common": "workspace:^",
     "@backstage/catalog-model": "^1.5.0",
     "@backstage/core-components": "^0.14.7",
@@ -43,7 +44,6 @@
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "4.0.0-alpha.61",
     "@types/react": "^16.13.1 || ^17.0.0 || ^18.0.0",
-    "qs": "^6.9.4",
     "react-use": "^17.2.4"
   },
   "devDependencies": {
